add, edit, display room information

+++ server/api/room/[id].delete.ts
@@ -6,14 +6,30 @@ export default defineEventHandler(async (event) => {
   const params = getRouterParams(event);
   const id = Number(params.id);
 
-  await prisma.room.delete({
+  const room = await prisma.room.findUnique({
     where: {
       id: id,
     },
   });
 
+  if (room) {
+    // Najpierw usuń pokój
+    await prisma.room.delete({
+      where: {
+        id: id,
+      },
+    });
+
+    // Następnie usuń sprzęt
+    await prisma.equipment.delete({
+      where: {
+        id: room.equipmentId,
+      },
+    });
+  }
+
   setResponseStatus(event, 202, "Accepted");
   return {
-    message: "room has been removed",
+    message: "room and associated equipment have been removed",
   };
 });
